fix(lookup): select text when reopening iframe
We keep the previous search populated in the lookup text field when
hiding the lookup iframe, so that when the user hits the hotkey again,
they can see their previous search.

However typically the user will want to start a fresh search, so
select all the text in the field, so that the user can start
immediately typing a new search without having to first type Control-a
or Cmd-a.

If instead they want to edit the previous search, they can press
cursor keys to deselect and start editing.