chore(background): make context menu click handler more robust
This handler is actually getting called not just when handling
clicks from the right-click context menu within a page, but also
when clicking the toggle item provided by webext-permission-toggle
within the extension context menu.

Since it's not suppose to handle anything in that case, make sure it
returns gracefully.

Also improve debugging in this handler function.