fix(manifest): make search hotkey work on all tabs
I had previously thought the `activeTab` permission was only in place
due to the documentation of the npm module
`webext-permission-toggle` (which rolod0x depends on) saying it was
required.

However, I had completely forgotten that it's a crucial part of the
mechanism for making the search hotkey work on any tab, rather than
just the tabs.

So restore the `activeTab` permission to the manifest, and document
why it's required in the relevant places.