docs(README): fix banner links on website
When links to other `.md` files are embedded as HTML, they work fine
on GitHub, but the HTML prevents the jekyll-relative-links plugin
from translating the URLs to end in `.html` on the website, so the
links break.

So turn them back into Markdown links, and add a Kramdown IAL to
center them.