feat(abbrev): add rough support for metasleuth.io
On metasleuth.io, addresses within the transaction graph are split
into two vertically stacked <text> elements within the <svg>: the top
one is 30+0, and the bottom is 0+10.  We'd need some custom code to
replace this elegantly, but for now arguably replacing just the top
one is better than nothing, even though it's ugly.