fix(options-storage): fix migration from v1 labels
The `OptionsSync.migrations.removeUnused` migration was running before
`migrateToSections`, which was removing labels before it could be
migrated into a new section.  So swap the order of migrations.

Also to be safe from defaults, use `isV1Options()` to check for the
presence of old labels and consider that as authoritative even if
there are already new sections.

Closes #500.